#2691 Moving integration tests to "kafka" package
