+++ integration-tests/kafka/src/test/java/io/quarkus/it/kafka/KafkaConsumerITCase.java
@@ -1,4 +1,4 @@
-package io.quarkus.it.main;
+package io.quarkus.it.kafka;
 
 import io.quarkus.test.junit.SubstrateTest;
