+++ integration-tests/kafka/src/test/java/io/quarkus/it/kafka/KafkaConsumerTest.java
@@ -1,4 +1,4 @@
-package io.quarkus.it.main;
+package io.quarkus.it.kafka;
 
 import java.util.Properties;
