+++ integration-tests/kafka/src/test/java/io/quarkus/it/kafka/KafkaProducerTest.java
@@ -1,4 +1,4 @@
-package io.quarkus.it.main;
+package io.quarkus.it.kafka;
 
 import java.time.Duration;
 import java.util.Collections;
